user registration

+++ admin/includes/view_all_posts.php
@@ -97,14 +97,13 @@ if (isset($_POST['checkBoxArray'])) {
                     echo "<td>{$cat_title}</td>";
                 }
                 echo "<td>$post_status</td>";
-                echo "<td><img width='150' src='../images/$post_image'></td>";
+                echo "<td><img width='100' src='../images/$post_image'></td>";
                 echo "<td>$post_tags</td>";
                 echo "<td>$post_comment_count</td>";
                 echo "<td>$post_date</td>";
                 echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
-                //<a href="../post.php?p_id=<?php echo $post_id;
                 echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
                 echo "</tr>";
             }
